Refactor view settings to use new BaseSettingsTab
Move view components to settings to avoid circular import
