+++ app/components/settings/__init__.py
@@ -18,3 +18,4 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 from .menu import SettingsMenu
+from .tabs import ViewContainer
