+++ app/components/settings/tabs/view/preview.py
